Fixed Shipping form Fixed validate line item method in order

+++ app/models/order.rb
@@ -240,9 +240,11 @@ class Order
   def validate_line_items
     line_items.each do |line_item|
       unless line_item.is_quick_item?
-        line_item.destroy unless line_item.item.present?
-        line_item.destroy if     line_item.item.ghost
-        line_item.destroy unless line_item.item.available
+        if line_item.item
+          binding.pry
+          line_item.destroy if     line_item.item.ghost
+          line_item.destroy unless line_item.item.available
+        end
       end
     end
   end
